refactor export modal
