+++ packages/geoview-core/public/locales/en/translation.json
@@ -46,7 +46,7 @@
     "opacity": "Opacity",
     "expand_legend": "Expand Legend",
     "toggle_cluster": "Toggle Clustering",
-    "toggle_opacity" : "Toggle Opacity"
+    "toggle_opacity": "Toggle Opacity"
   },
   "datagrid": {
     "exportJson": "Download GeoJSON"
@@ -70,5 +70,10 @@
       "createtwice": "Can not execute twice the createGeoViewRasterLayers method for the map __param__",
       "usedtwice": "Duplicate use of geoview layer identifier [__param__] on map __param__"
     }
+  },
+  "exportModal": {
+    "title": "Export Map as PNG",
+    "cancelBtn": "Cancel",
+    "exportBtn": "Export"
   }
 }
