+++ packages/geoview-core/src/core/components/export-modal/export-modal.tsx
@@ -1,5 +1,5 @@
 import { MouseEventHandler, useContext } from 'react';
-
+import { useTranslation } from 'react-i18next';
 import { Button, Dialog, DialogActions, DialogTitle } from '@mui/material';
 
 import { MapContext } from '../../app-start';
@@ -28,12 +28,11 @@ const defaultProps = {
  */
 export default function ExportModal(props: ExportModalProps): JSX.Element {
   const { className, isShown, closeModal } = props;
-
+  const { t } = useTranslation();
   const mapConfig = useContext(MapContext);
   const { mapId } = mapConfig;
 
   return (
-    // eslint-disable-next-line react/jsx-no-bind
     <Dialog open={isShown} onClose={closeModal} className={className}>
       <form
         onSubmit={(e) => {
@@ -41,13 +40,13 @@ export default function ExportModal(props: ExportModalProps): JSX.Element {
           exportPNG(mapId);
         }}
       >
-        <DialogTitle>Export Map as PNG</DialogTitle>
+        <DialogTitle>{t('exportModal.title')}</DialogTitle>
         <DialogActions>
           <Button onClick={closeModal} size="small" autoFocus>
-            Cancel
+            {t('exportModal.cancelBtn')}
           </Button>
           <Button type="submit" onClick={closeModal} size="small">
-            Export
+            {t('exportModal.exportBtn')}
           </Button>
         </DialogActions>
       </form>
